# src/main/resources/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Studeal API",
    "description": "Studeal의 API 명세서 🌱",
    "contact": {
      "name": "Studeal Team"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/enrollments": {
      "post": {
        "tags": [
          "수업 참여 확정"
        ],
        "summary": "최고가를 부른 학생과 선생님을 연결하는 API",
        "description": "경매 종료 후, 최고가를 부른 학생과 선생님을 연결하는 API입니다.이를 통해 학생은 낙찰받은 수업을 결제하여 거래를 성사하거나, 거절하여 최종 낙찰을 포기할 수 있습니다.",
        "operationId": "createEnrollment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.studeal.team.domain.enrollment.dto.EnrollmentRequestDTO$CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "NEGOTIATION404": {
            "description": "존재하지 않는 협상입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          },
          "STUDENT4001": {
            "description": "존재하지 않는 학생입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          },
          "ENROLLMENT4001": {
            "description": "이미 Enrollment(학생의 수업 참여 확정 현황)가 존재합니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          },
          "ENROLLMENT4003": {
            "description": "협상이 성공 상태가 아닙니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/signup": {
      "post": {
        "tags": [
          "사용자 인증"
        ],
        "summary": "회원가입 API",
        "description": "사용자 회원가입을 처리하는 API입니다. 학생/강사 역할에 따라 처리됩니다.",
        "operationId": "signupUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.studeal.team.domain.user.dto.UserRequestDTO$SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "USER4003": {
            "description": "유효하지 않은 사용자 역할입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          },
          "COMMON200": {
            "description": "OK, 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponseCom.studeal.team.domain.user.dto.UserResponseDTO"
                }
              }
            }
          },
          "USER4002": {
            "description": "이미 등록된 이메일입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          },
          "USER4004": {
            "description": "유효하지 않은 비밀번호 형식입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          },
          "USER4005": {
            "description": "이름은 필수입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          },
          "USER4006": {
            "description": "유효한 이메일 형식이 아닙니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": [
          "사용자 인증"
        ],
        "summary": "로그인 API",
        "description": "이메일과 비밀번호로 로그인하여 JWT 토큰을 발급받습니다.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "COMMON200": {
            "description": "OK, 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponseCom.studeal.team.domain.user.dto.TokenDTO"
                }
              }
            }
          },
          "USER4001": {
            "description": "잘못된 인증 정보입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/enrollments/{enrollmentId}/status": {
      "patch": {
        "tags": [
          "수업 참여 확정"
        ],
        "summary": "수업 참여 확정 변경 API",
        "description": "수업 참여 확정 상태를 변경하는 API입니다. (WAITING → CONFIRMED/CANCELED, CONFIRMED → CANCELED 변경 가능) 이 API를 통해 학생은 수업 참여 확정 상태를 업데이트할 수 있습니다.",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.studeal.team.domain.enrollment.dto.EnrollmentRequestDTO$StatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "ENROLLMENT404": {
            "description": "존재하지 않는 수강 신청입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          },
          "ENROLLMENT4002": {
            "description": "유효하지 않은 상태 변경 요청입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.studeal.team.global.error.ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "com.studeal.team.global.error.ApiResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "description": "성공 여부",
            "example": true
          },
          "code": {
            "type": "string",
            "description": "응답 코드",
            "example": "COMMON200"
          },
          "message": {
            "type": "string",
            "description": "응답 메시지",
            "example": "성공입니다."
          },
          "result": {
            "type": "object",
            "description": "응답 데이터"
          }
        },
        "description": "API 응답 형식"
      },
...
      "com.studeal.team.domain.enrollment.dto.EnrollmentRequestDTO$CreateRequest": {
        "required": [
          "negotiationId",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "negotiationId": {
            "type": "integer",
            "format": "int64"
          },
          "studentId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "com.studeal.team.domain.user.dto.UserRequestDTO$SignupRequest": {
        "required": [
          "email",
          "name",
          "password",
          "role"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "사용자 이름",
            "example": "홍길동"
          },
          "email": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "description": "이메일",
            "example": "user@example.com"
          },
          "password": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "description": "비밀번호",
            "example": "StrongPassword1!"
          },
          "role": {
            "type": "string",
            "description": "사용자 역할",
            "example": "STUDENT",
            "enum": [
              "STUDENT",
              "TEACHER"
            ]
          },
          "bio": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string",
            "description": "자기소개",
            "example": "안녕하세요, 대학생 홍길동입니다."
          },
          "major": {
            "type": "string",
            "description": "전공 과목",
            "example": "MATH",
            "enum": [
              "MATH",
              "SCIENCE",
              "ENGLISH",
              "KOREAN",
              "HISTORY",
              "CODING"
            ]
          }
        },
        "description": "회원가입 요청"
      },
      "com.studeal.team.domain.user.dto.UserResponseDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "STUDENT",
              "TEACHER"
            ]
          },
          "bio": {
            "type": "string"
          },
          "major": {
            "type": "string",
            "enum": [
              "MATH",
              "SCIENCE",
              "ENGLISH",
              "KOREAN",
              "HISTORY",
              "CODING"
            ]
          }
        },
        "description": "응답 데이터"
      },
      "com.studeal.team.global.error.ApiResponseCom.studeal.team.domain.user.dto.UserResponseDTO": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "description": "성공 여부",
            "example": true
          },
          "code": {
            "type": "string",
            "description": "응답 코드",
            "example": "COMMON200"
          },
          "message": {
            "type": "string",
            "description": "응답 메시지",
            "example": "성공입니다."
          },
          "result": {
            "$ref": "#/components/schemas/com.studeal.team.domain.user.dto.UserResponseDTO"
          }
        },
        "description": "API 응답 형식"
      },
...
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "이메일",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "example": "StrongPassword1!"
          }
        },
        "description": "로그인 요청"
      },
      "com.studeal.team.domain.user.dto.TokenDTO": {
        "type": "object",
        "properties": {
          "grantType": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessTokenExpiresIn": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "응답 데이터"
      },
      "com.studeal.team.global.error.ApiResponseCom.studeal.team.domain.user.dto.TokenDTO": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "description": "성공 여부",
            "example": true
          },
          "code": {
            "type": "string",
            "description": "응답 코드",
            "example": "COMMON200"
          },
          "message": {
            "type": "string",
            "description": "응답 메시지",
            "example": "성공입니다."
          },
          "result": {
            "$ref": "#/components/schemas/com.studeal.team.domain.user.dto.TokenDTO"
          }
        },
        "description": "API 응답 형식"
      },
      "com.studeal.team.domain.enrollment.dto.EnrollmentRequestDTO$StatusUpdateRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "WAITING",
              "CONFIRMED",
              "CANCELED"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT 인증 토큰을 입력해주세요. \u0027Bearer [토큰]\u0027 형식으로 입력",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}